Parsed a JSON array
We are checking in this code prior to some code cleanup, if nothing
else, so we can to celebrate!

This is a checkpoint at the point where the compound JSON array
has been successfully parsed, even when JSON arrays are
nested in JSON arrays:

    ./jparse -J 3 -s '[1,[10,20,[111,222,333]],[100,200,[1000,2000,3000,4000]]]'

Added JTYPE_ELEMENTS, a type of JSON node that is needed to hold values
within a JSON array as they are being parsed and BEFORE the JSON array is ready.

Renamed "enum element_type type" to "enum item_type type".
Renamed the union in "struct json" from "elements" to "items".
Added 'struct json_elements' to the union.

   This had to be some because we parse JTYPE_ELEMENTS while the JSON
   array is being formed.  We want to avoid confusion between the generic
   term "element" and the JSON element(s).

Comments that used the term "JSON element", that were NOT referring to an
actual JSON element or JSON elements, have been changed to "JSON item".
Well mostly .. there are probably a number of cases that we missed
and/or incorrectly modified.

The general hack to parse JSON arrays was to build up (using a
dynamic array of pointers to JSON values) a JSON node of type
JTYPE_ELEMENTS.  When the JSON array is ready to be parsed, the
type JTYPE_ELEMENTS is converted into type JTYPE_ARRAY.  However,
until the JSON array is parsed, the parser continues to add JSON
values to the JSON node of type JTYPE_ELEMENTS.

The above complication (JSON node of type JTYPE_ELEMENTS) was needed
because a JSON array have have as one of its JSON values, a lower
level JSON array.  See the above jparse command for an example.

There is some code that of blocked form compiling with:

    #if 0 /* XXX - code likely not needed - remove this section if that is the case - XXX */
    ...
    #endif /* XXX - code likely not needed - remove this section if that is the case - XXX */

Such code is very likely to be removed soon.

Sequenced exit codes.

TODO: Test more JSON array parse cases and add them to json_teststr.txt.

TODO: Fix a number of comments and error messages that should / need to be changed
      or removed as a result of the above changes.

TODO: Tune JSON debug levels.

    For example, at -J 3 we do not need to dump JSON trees at every parse stage.

TODO: Parse JSON object using lessons learned form compound JSON array / JSON elements parsing.

TOOD: Make use of dbg_allowed(level) in dyn_array.c to reduce the CPU impact of
      dbg() calls that lots of args.

TODO: Improve ugly_error() messages.

    It is very possible that ugly_error() in jparse.y should be improved
    to allow for reporting filename and line number of an error.
    When that is done, then ugly_error() can be improved to print such
    parser information using parser values.

    The ugly_error() function should NOT call the chk_foo() code
    when this is done.  Let the pure JSON parser use just ugly_error(),
    parser internal values along with the (currently unused) jwarn(),
    jwarnp(), jerr(), and jerrp() functions (perhaps modified as needed).