Parse args in hostchk.sh
The usage is now:

    $ ./hostchk.sh -h
    usage: ./hostchk.sh [-h] [-v level] [-D dbg_level] [-t tar] [-c cc]

	-h			    print help and exit 2
	-v level		    set verbosity level for this script: (def level: 0)
	-D dbg_level	    set verbosity level for tests (def: level: 0)
	-t tar		    path to tar that accepts -J option (def: /usr/bin/tar)
	-c cc		    path to compiler (def: /usr/bin/cc)

    exit codes:
	0 - all is well
	1 - at least one test failed
	2 - help mode exit
	3 - invalid command line
	>= 30 - internal error

Notice that the default paths to tar and cc are set up by type -P prior
to setting up to the USAGE variable string. Previously tar was always
/usr/bin/tar and cc was just cc by itself (no path specified).

Add firewall checks for tar and cc and update references to these
commands to be "$TAR" and "$CC" instead of tar and cc.

Much more has to be added to this script as well as (the still
non-existent) bug-report.sh script.