Prepare for chkentry, added chkentry stub code
The chkinfo and chkauth tools will be replaced by chkentry.
References to chkinfo and chkauth have been replaced with chkentry.
This commit puts in place, a stub for the chkentry command:
a wrapper of a program that does nothing (yet).

Files of the form chk* have been replaced to  old.chk* and
are no longer compiled / references by Makefile.
The old.chk* files will set around for a bit in case
there is useful code / text that needs to be copied from them.

The flags for chkentry are just a best guess and may be changed as
code for chkentry is added.  The mkiocccentry current uses -C
to specify the path to chkentry.  This letter was picked as a
best guess and may be changed as code for chkentry is added.

A stub chkentry.1 man page was added.

In .info.json the chkinfo_version was replaced by chkentry_version.

In .author.json the chkauth_version was replaced by chkentry_version.

The IOCCC_info_version was changed from 1.9 2022-03-15 to 1.10
2022-06-22.  The INFO_VERSION was changed accordingly.

The IOCCC_author_version was changed from 1.12 2022-03-15 to 1.13
2022-06-22.  The AUTHOR_VERSION was changed accordingly.

The CHKINFO_VERSION and CHKAUTH_VERSION have been replaced by
CHKENTRY_VERSION, which current stands at "0.0 2022-06-22"
(to reflect its stub code status).

The MIN_TIMESTAMP has was changed from 1652598666 to 1655958810.

A number of test_txzchk files were renamed with numeric
values in the 192xxxxxxx range to keep them well above the
MIN_TIMESTAMP.