Updates to chkentry.c and chkentry.h
Remove unused strict boolean. As I recall this was an obsolete idea and
it was not settable via the command line either so there was no need to
have it.

I also changed the checks for '== false' and '== true' and changed the
more than one call to exit to be simply

     exit(error != false);

As for comments I changed the header to refer to me working on it as it
had been with chkinfo.c and chkauth.c. This could be changed of course
but it seemed fitting.

One of the next things I'll be working on with this file (most likely)
is adding a call to the find_utils() function and then parsing the
required args (as well as adding the proper sanity check function).
These can come from the now called old.chk{info,auth}.c. There probably
are other functions that have to be moved or copied in full or part. For
example the check functions (the ones that check the json values) could
be used to save having to rewrite some of the same code (the only real
difference is the tree itself so how it's traversed will be different).