jparse_test.sh typo fix and behaviour modification
In the usage message the -D was described twice.

Instead of cognitive behavioural modification (CBT) we now have script
behavioural modification (SBT) :) as follows: The script now reads by
default the test suite text file json_teststr.txt. To read from stdin
use the Unix de facto standard -.

There is a reason for the above change in functionality: often I mean to
run the default tests but it blocks as it reads from stdin. Then I
remember that it default reads from stdin. However the test cases we
have used as we were developing the parser are in json_teststr.txt so
this seems a better default file to use. Since - signifies stdin in many
tools we use that for reading of stdin instead of defaulting to stdin.