Make jparse.1 consistent with recent developments
Removed from FILES section files that no longer exist.

Make the description of sorry.tm.ca.h more clear (it still could be
improved: it can be argued there is a bit of redundancy but this will
have to do for now).

Update BUGS to note that although jparse is in a sense finished the
parser does not return a tree (even though everything is parsed). I also
noted that better error reporting and various other things (those words
exactly as there are things that are not yet determined) need to be
added.

Added AUTHORS section.