JSON parse functions don't object to only spaces
Instead the scanner and parser will deal with this. Thus instead of:

    $ printf '\0' | ./jparse -
    Warning: parse_json_file: stdin is empty
    invalid JSON

it is now:

    $ printf '\0' | ./jparse -
    Warning: ugly_lex:
    invalid token: 0x00 = <>
    in ugly_error(): node type: ((node == NULL))
    JSON parser error on line 1: syntax error
    invalid JSON

This is because the reason it is invalid is the NUL byte and not because
it's all spaces or a NUL byte. NUL bytes are allowed in strings in JSON
and although in this case that's not relevant we want to report the
correct reason it is invalid JSON.