dbg interface documentation updates  …
Update the dbg.3 man page to refer to the functions added in commit
09b6563:

    extern void snmsg(char *str, size_t size, char const *fmt, ...)
    extern void sndbg(char *str, size_t size, int level, char const *fmt, ...);
    extern void snwarn(char *str, size_t size, char const *name, char const *fmt, ...);
    extern void snwarnp(char *str, size_t size, char const *name, char const *fmt, ...);
    extern void snwerr(int error_code, char *str, size_t size, char const *name, char const *fmt, ...);
    extern void snwerrp(int error_code, char *str, size_t size, char const *name, char const *fmt, ...);

Additionally the dbg_example.c file with a message for each output that
says what the next line should print so that the user can verify it's
correct. It does not make use of the sn functions above (though I did
think of this to do a strcmp()) as it's just a simple example program.
The source code has been updated in the man page and the output has as
well.