Typo fixes in find_utils()
When an alternate location was found to be executable it said it would
try using the location but it actually will use it. There's an argument
that it could fail in that the correct options do not exist so it might
not succeed but even so it's using it whether it fails or not so it's
not trying to use it.