Update JSON parser version to 0.10 2022-07-03
The previous version was 0.8 2022-06-12 but since that time the parser
was code complete (this happened with commit
7f8d7411c9157fe314a237a3eec5c404a4de3c79 at Thu Jun 23 15:41:54 2022
-0700) which by itself warrants a new version. That should be 0.9
2022-06-23.

But earlier this morning (00:48:17) some additional changes were made
and since the jparse version was updated at this point I believe it
worth updating the json parser version as well. The two now are the same
version: 0.10 2022-07-03.