Minor fixes in jparse filename check
Previously the check was if the file was a file but it reported it as if
did not even exist. Now check first if it exists and report that it does
not exist if it does not exist. Then check if it's a normal file and if
it's not report that it's not a normal file.