JSON parser now uses werr() instead of warn()
Technically these are errors but we don't want them to be fatal so we
use werr() instead of warn(). This could be changed back if this is not
actually desired. Where errno is relevant use werrp() instead of werr().

I also added a couple XXX comments about whether certain conditions
should actually be errors (internal errors in the parser which for now
are also just non-fatal errors).

Updated comments about warning as well as removing the part that says
it (issue warnings instead of errors) is subject to change as the parser
is developed: as the parser is code complete (though some things still
have to be added) and it's been decided that it should not be fatal so
that a full report can be given (if requested).