Move { to line below struct in header files
Some structs still had the form

    struct name {
    };

instead of

    struct name
    {
    };

and the latter form was agreed to be the preferred form (much like with
functions).

Note that there are still some cases in some C files (in initialisation
of arrays of structs) where the { is on the same line as the 'struct'
but I'm not sure if these should be changed or not. Probably yes but I'm
leaving them as is for now.