Change 'being worked on' to 'co-developed by'
There were several files where the comments stated that the parser was
being worked on by me but they should have been updated to say it was
co-developed by Landon Curt Noll and me which they now do.

Additionally in jsemtblgen.c I added this comment like it is in
jsemtblgen.h so that it matches.