Make use of print() macro for printing versions
In some files the errno reset was removed as the print() macro does this
(via pr() -> vfpr()) itself for each action (writing the string and
flushing the output).

For some others the macro is now used instead of printf().

In mkiocccentry printing the welcome string is no longer fatal and it
uses the print() macro as well. It doesn't appear to be fatal in any
other tool (I know it wasn't fatal in txzchk but I verified it when I
changed it to also use the print() macro).