Minor updates to jparse.1
Add to FILES section jparse_main.h and jparse_main.c.

Note that jparse_test.sh is to test jparse itself (even if this should
be obvious it was not made clear in the previous version: it could have
been a way to read more than one blob of json but the point is to test
that the parser works rather than to just parse one or more blobs of
json).