Fix checknr in Makefile + typo fix dbg.3
Add -c option to checknr invocation to add commands that checknr does
not know about but which are valid commands.

Removed a .RE that had no matching .RS (and did not need one) in dbg.3.

Now make checknr is clean (previously only section 1 of the man pages in
the repo were checked but as of commit 04df8796 that is no longer true).