txzchk no longer links in json related objects
As of commit f5caa919547e762053247cfb7479dc8d51eb7c13 on 16 June 2022
sanity checks in sanity.o do not need any json code but the Makefile was
not updated to not link in json_parse.o and json_util.o. Originally I
added json checks to it to be a more comprehensive check on the entire
toolkit (which is why txzchk needed certain json object files) but since
this is no longer the case there is no need to recompile txzchk when
these object files change.

Additionally I added some humour to the comments. Since txzchk checks
for problems with the tarball the reference to 'tar and feather' refers
to these problems (where the problems are the feathers).

I also moved the 'Because sometimes even the IOCCC Judges need some
help' so that it's under my name as it flows better (i.e. 'it was
written by me because sometimes even the IOCCC Judges need some help').

The comment changes were added to txzchk.{c,h} and txzchk_test.sh.