More fun with tar and feathers :-)
Instead of saying 'All checks passed.' txzchk now says 'No feathers
found in tarball.' if all tests pass.

Also the top comment now says more about what txzchk actually does (it
did not even state that it verifies that the tarball conforms to the
IOCCC tarball rules as it now does) but having fun with the idea of the
problems checked for are feathers.