Share and vomit on the JSON spec! :-)
As a play on the 'Share and enjoy! :-)' that was at the top of some of
the JSON files they now all (those that said it and those that did not)
say:

    + * Share and vomit on the JSON spec! :-)

This is because it's an emetic :) and it's just more fun with what
Landon and I have been having by satirising JSON for the IOCCC.

Additionally for the JSON files that did not state that the parser was
co-developed by Landon and me they now say it. I also added to all the
files (that were missing it):

    + * "Because sometimes even the IOCCC Judges need some help." :-)

Run make rebuild_jnum_test parser all.