Improve type -P usage in scripts
This expands on commit e212ffd0 which did two things I did not think of
in jsemcgen.sh:

It redirected stderr of the type -P command to /dev/null and it also
checked if it was empty (I did not even think of the fact it could have
been empty).

These are good but it can be made better by checking if it is empty
prior to the usage message and (if empty) set it to the hard code
default (directory being /usr/bin in all cases).

We cannot do it via parameter substitution because it attempts to
execute the command. In the case of tar and cc this is not much of a
problem: we can simply redirect stderr to /dev/null. But in the case of
patch it blocks (running patch without any args is not an error) so we
have to do it by checking for an empty variable and then assign it
specifically. To be safe we do this for cc and tar as well.

The scripts hostchk.sh (which is still incomplete), jsemcgen.sh and
txzchk_test.sh (which still has to be improved / is incomplete) have
been updated for this. The vermod.sh script has not been updated because
it does not allow one to pass an alternate path to rpl (maybe it should
be allowed but it does not currently so I've not touched it).