Improve Makefile wrt invocation of shell commands
Instead of using backticks / grave accents use built-in make 'shell'
function. At first I thought to use '$$(...)' but 'shell' seems like it
will be more portable. This has been done with all cases of backticks in
the Makefile and I have run the various rules to `make sure' that it
makes everything okay.