Update dbg.3 for 7941077df functions + other fixes
In addition to the new functions from commit 7941077df have been added
to the man page I have moved the errno versions higher up and also made
a few fixes and clarifications.

The dbg.md file has to be updated still but this will very possibly not
be done today as it will take more time and thought.