Change header comments in chkentry.{c,h}
Now refers to the JSON parser being developed by Landon and me as well
as the tool being co-developed by us and the concept of the file by
Landon.

A subtle point on the json parser. I changed the 'this' to 'the' because
chkentry is not actually the parser; it uses the parser but it's not the
parser. There might be other files that have this problem as well.