Add check for bogus fields in txzchk
In other words if there's a field after the filename it's an issue. This
might be a file with a space in the name. This won't trigger invalid
filename (as the strtok_r() will have removed the space) but the tarball
will now report having some feathers in it. This has been done for both
linux and BSD/macOS formats.