Check for multiple bogus fields in txzchk
In other words a line like:

    -rw-r--r--  0 501    20          0 Jul 28 08:25 test-0/file bogus data

will report (for this line specifically):

    Warning: txzchk: test_txzchk/entry.test-0.1658587455.txt: bogus field found after filename: bogus
    Warning: txzchk: test_txzchk/entry.test-0.1658587455.txt: bogus field found after filename: data

Previously it would have only detected 'bogus'.

A question that was brought up is whether or not the line should count
as a filename of 'file bogus data' or if considering 'bogus' and 'data'
as bogus fields is enough. On the one hand the most likely case is it
would be a filename with spaces (unless of course the tarball was
created by mkiocccentry); but on the other hand it might also be someone
messing with the tarball and since strtok_r() is used (with " \t") it
would require more code to detect and display the full filename and
since it's also bogus this works (for now if not indefinitely).