Add test_extra_file() to entry_util.{c,h}
The function checks that the filename length is <= MAX_BASENAME_LEN and
that it's POSIX portable filename and + chars i.e. matches the regexp:

    ^[0-9A-Za-z][0-9A-Za-z._+-]*$

The function is not called yet.

Also changed the message for empty abstract. It now says:

    invalid abstract: is empty

whereas before it said:

    invalid empty abstract

which could be read as it could be valid in some cases (at least when
looking at the message only). In any case this specifically states the
problem.

For the extra_file filename being 0 length it says that it's 0 length
because there are other tests that are done on it.