Minor updates and fixes in txzchk
Change remaining references of 'issues' to 'feathers' to continue the
fun of:

    "Because sometimes people add feathers to tar." :-(

(There are still a few references to issues but this is only to make it
clear that it's just wordplay: all messages and variables now refer to
feathers though.)

I've also improved show_txz_info() in the following ways:

It now shows the booleans as true or false.

I have attempted (keyword) to make it all the debug information lined up
(matching columns for names and values). This might not always work as
\t is funny like that but for some tarballs it works fine at least. For
example with the file test_work/entry.test-0.1659097955.txz the
alignment is still not consistent for every line but for
test_txzchk/entry.test-0.1658587455.txt (which is only on my system) it
works fine. It might be a bit better though.

I thought of making this a new version but that would require updating
the version in the json files and the changes are only cosmetic anyway
so it doesn't warrant a new version anyway.