Improved & fixed struct info & struct author, added struct auth
Converted struct author to only refer to an IOCCC author array fields.
Removed common element form author as the structure was duplicating
data for each author that was not related to a specific IOCCC author.

Created a struct auth to hold the IOCCC .info.json information.
Set struct auth field order to refer to the canonical order
of the .auth.json file as written by mkiocccentry.

Simplified the structures for the IOCCC .author.json information
and IOCCC .info.json information by eliminating struct json_common.
Simplified code that used the common sub-structure.

Changed order of struct info to refer to the canonical order
of the .info.json file as written by mkiocccentry.

Added no_comment as a compiled in in constant to info & auth.
Added author_version as a compiled in in constant to auth.
Added ioccc_contest as a compiled in in constant to info & auth.
Added info_file as a compiled in in constant to info.
Added author_file as a compiled in in constant to info.

Made mkiocccentry_ver into a compiled in constant.
Made epoch into a compiled in constant.

Changed field strings that are compiled in to 'char const *'

Fixed how free_author_array() to zeroizes each array member.

Updated free_info() to free all non const strings.
Changed order frees in free_info() to to match struct info order.
Added comments about compiled in constant that cannot be freed.

Added free_auth() to entry_util.c to free struct auth.

Made use of free_auth() in mkiocccentry.c.
Made use of struct auth in mkiocccentry.c.

Sequenced exit codes.