Typo fixes in util.{h,c}
Function fprint_line_buf() referred to the no longer existing function
fprint_str() in a comment in a way that suggests it still exists. Update
the comment to say that it's an old function.

Also the macro that fprint_str() used (USE_STRLEN) is no longer in use
so I have removed it.