Add foo
Q: The question is what should the file be called? It could be that.
But it could be something else. What do you think? The chkentry.h
file will have to include the header file. I could move the function
to the source file as well and link it in.

A: foo :-)