Fix file size conversion in txzchk
In parse_bsd_txz_line() if errno != 0 after call to strtoll() (now
strtoimax() like in json_parse.c) it did not increase the count of
feathers (issues) and neither did it test for size < 0. It now does
both.

Additionally if errno != 0 it (in the linux function too) after
increasing feathers and warning about it it returns as there's nothing
else it can do with that line. It is not an error with conversion but
perhaps it should. If that's ever the case then a new function say
string_to_off_t could be added to util.c but I'm not sure if this is
strictly necessary.

What is not yet tested is the file size variables being >= OFF_MAX as
I'm not sure all the tests that should be added (this will be discussed
and if necessary addressed).

These bugs were discovered when I was making the formatting my typical
style ({ and } on lines by themselves as well as not having {} pairs for
one line bodies).