Testing legacy / timegm detection and workarounds
Renamed legacy_os to have_timegm in this repo.

Removed ${LEGACY_CFLAGS}.  By default ${CFLAGS} is used.

By default, the -lm is not used and ${LDFLAGS} is empty.

By default, ${D_LEGACY} is empty.

By default, ${WARN_FLAGS} is -Wall -Wextra -Werror only.

Added code in Makefile to run have_timegm.sh in order to see if the
system can compile the c11 function timegm().  It if cannot, then
${TIMEGM_PROBLEM} will be set to -DTIMEGM_PROBLEM.  This is turn
will cause this to be added to ${D_LEGACY}:

    -D_DEFAULT_SOURCE -D_ISOC99_SOURCE -D_POSIX_C_SOURCE=200809L -D_XOPEN_SOURCE=600 -DTIMEGM_PROBLEM

and this is added to ${LDFLAGS}:

    -lm

and this is added to ${WARN_FLAGS):

    -Wno-unused-command-line-argument -Wno-poison-system-directories -Wno-unreachable-code-break -Wno-padded

When TIMEGM_PROBLEM is fined, entry_util.c adds a extern declaration for timegm().

Added language to the Makefile and entry_util.c explaining the
work-a-round / gross hack supporting pre-c11 systems such as Centos 7:
that such systems are supported on a best guess basis and support
will go away in 2024.