Fix 16beee5 for CentOS compilation
Because chkentry.c uses a function that is defined in foo.c chkentry.h
included foo.h. However this resulted in a compilation error in CentOS
as supposedly an array that was defined was not being used (this is
because it's used in foo.c but not chkentry.c but chkentry.c #included
foo.h for above mentioned function) i.e. it was used in one place but
not the other.  Removing static would not work because then it would
result in multiple definitions of the same symbol.

There were three fixes I could think of. One was discarded before as
undesirable (inclusion guards to enable certain variables) and another
would be a crime against foo so I opted with moving the function
declaration into chkentry.h since that's where it's used anyway.

Run make depend and fix typos in Makefile.