Several fixes and enhancements to txzchk
The following fixes were made (but note that if a test is not mentioned
below it does not mean it's not already in place and it's entirely
possible I miss below one or more tests that has or have been added).

After each file size is parsed if the new total size is < the previous
size it is an issue.

After each file size is parsed if the new total size is > MAX_DIR_KSIZE
it is an issue (this was only done at the end after all files parsed).

After each file size is parsed if the new total size is < 0 it is an
issue (the current file size was tested for < 0 and also 0 but the total
was not until the very end).

After each file size is parsed if the new rounded file size (up to 1024
multiple) is < the previous it is an issue.

Each test is only displayed once. That is there are new booleans that
disable showing a warning and also increasing the number of feathers
(issues) in the tarball but perhaps the total feathers should be
increased and only the message should not be. The reason I did it this
way is because it's simpler to keep track of the values and if for
example a 0 file size comes in between what should the message be?
Another reason I opted for this is it's less verbose as the tool is
already quite verbose without appropriate options (-q or -e but the
latter should be removed and will be eventually when the test script is
updated to show more details).

Perhaps all the checks made on the tarball should be added to the man
page but if so that's for another time.

I have not added any of these to the final report but perhaps they
should be.

Perhaps the booleans should be an unsigned int being the number of times
it occurs?