More test_ functions added + typo fixes
The following functions were added:

    extern bool test_name(char *str);
    extern bool test_no_comment(char *str);
    extern bool test_past_winner(bool boolean);

I was going to add the two rule size test functions but I'm unclear
about something so I'll be posting a comment instead to discuss it once
a pull request has been formed.

Typo fixes in entry_util.c and mkiocccentry.c. This includes changing
pseudo-name to pseudonym - this fix could possibly cause a logomachy but
I deem that risk worth it as it should be correct! :-)