Release new txzchk version - 0.11 2022-08-23
The new version is due to many commits since the last version update
including bug fixes and enhancements. In this commit in particular:

The unsigned variables are now all uintmax_t.

Count the number of times total files size (and rounded total files
size) go below 0, become 0 and become > the max.

There was also a bug fix: a potential file descriptor leak because the
pipe from popen() (or in the case of -T the regular file from fopen())
was not closed until after parsing all lines but a number of functions
exit on error so this would never be done (even if the OS will do it
it's better to do it first).

The txzchk_version in the test_JSON files has been updated to the new
version.

There might have been other changes as well that I'm forgetting.

One thing that should possibly be done is reword the debug messages in
show_txz_info(). Right now because of different length strings the tabs
will cause unaligned text. Thus lines like:

    debug[3]: txzchk: entry.test-1.1660750765.txt: has .info.json:		true
    debug[5]: txzchk: entry.test-1.1660750765.txt: empty .info.json:		false
    [...]
    debug[5]: txzchk: entry.test-1.1660750765.txt: times size of all files rounded up to multiple of 1024 shrunk:		1
    debug[5]: txzchk: entry.test-1.1660750765.txt: times size of all files rounded up to multiple of 1024 went below 0:		0
    debug[5]: txzchk: entry.test-1.1660750765.txt: times size of all files rounded up to multiple of 1024 went above max 27651:		3

could actually be reworded so that it's a normal sentence e.g.:

    debug[3]: txzchk: entry.test-1.1660750765.txt: has a .info.json file
    debug[3]: txzchk: entry.test-1.1660750765.txt: does not have an empty .info.json file
    debug[5]: txzchk: entry.test-1.1660750765.txt: the total files size rounded up to 1024 file size shrunk 1 time

though even this is not perfect and maybe is not worth it: not sure.
Maybe also it should not report the counts if the condition never
happened. These can be decided later or not.