Programmatically regenerate oebxergfB.h
It's now possible for me to create this file without having to copy and
paste anything.

This does involve updating .gitignore as there's a text file that cannot
ever be committed.

A TODO would be to write a script or Makefile rule that does this to
prevent accidental overwriting the wrong file (though this is not really
a problem as I can always redo it quickly but it would still be nice to
have).

Added several strings to oebxergfB.h at the same time.