Verify .info.json, .author.json and fnamchk exist
The code now verifies that the .info.json (if not excluded from check),
.author.json (if not excluded from check) are regular files and can be
read and that fnamchk can be read and is executable. The following
functions have been added:

    static void chkentry_sanity_chks(char const *entry_dir,
	    char const *info_json, char const *author_json,
	    char const *fnamchk);
    static void verify_entry_dir(char const *entry_dir,
	    char const *info_json, char const *author_json);

There IS an issue to resolve but this must wait for other functions to
exist which have not yet been written (and on my behalf I haven't even
thought about it - yet). This is noted with a couple XXX comments.
Specifically the function verify_entry_dir() is designed to prevent code
duplication but at this time it would actually (if not modified) create
code duplication.

The function verifies that the correct files exist (if the entry_dir ==
NULL then it's expected that .info.json and .author.json paths have been
specified where a '.' means ignore that file). It does this (in the case
of a directory being specified which is the single arg form of the
command) by using chdir(2) (first recording the current working
directory with open(2) with path ".") and then (if it succeeds) it runs
the usual exists(), is_file() etc. functions on them. If the directory
!= NULL then we specifically set the filenames to ".info.json" and
".author.json"; if it's NULL the user will have specified the paths.

Now as for the duplication the problem is that we will later have to
chdir() into the directory in that form which will be a duplication of
the code. There are at least two ways to solve this problem which I have
noted. One is that there are functions that get the current working
directory and then change into the other directory and one that also
undoes these actions. Whether these belong in util.c or not I don't
know. Another option I thought of is that we only check utility paths
(or at this point single path - fnamchk) in the chkentry_sanity_chks()
function and let the functions that process the files (or directory) do
the checking. However this could also result (and I only just thought of
this) code duplication depending on how it's done as one way to get the
files is by directory and another is direct paths. But since the
functions that actually validate the files have not been created yet
these issues can be answered later.

One minor thing about this function is it might be a misnomer or
misleading: it does not actually validate that the directory's
.info.json and .author.json files are correct but rather that the files
exist. Perhaps a better name could be devised but for now this will do.

As for chkentry_sanity_chks() this function verifies that fnamchk can be
found (we do use find_utils() first) and it also calls the
verify_entry_dir() done carefully (just like the function itself) so
that both forms of the command can work (which means that in one form
the info_json and author_json should be NULL and the other the entry_dir
should be NULL but all three cannot be NULL). There's no need to check
for one (but not both) of the info_json and author_json files being NULL
because when only one arg is specified it's the directory form and the
filenames are set to NULL.