Fix commit a276c44e606bc7f2a470d0a6805904c88
This commit divorces the verify_entry_dir (now called
validate_entry_files()) from chkentry_sanity_chks(). If
chkentry_sanity_chks() does not pass the validate_entry_files() will not
run. If it does pass it will run and depending on what args are used and
what is readable it will run (for now only stub) functions to validate
the respective files.

What this means is that for now if the file is invalid json the program
will return exit code of non-zero. If the json is valid it will be 0.
This addresses the concern of possible code duplication.

This involved updating the Makefile.