Update dbg facility documentation
The dbg.md file has been dramatically reduced in size, referring the
user to the man page instead. The man page has been updated with some
notes that were only in the md file and which is no longer there.

I will be doing a commit in the dbg repo and making a pull request there
too. Should anything need to be fixed I can do that later either before
or after a merge.