Typo fixes + formatting + rewording CHANGES.md
There's some humour in there as well: literally as well as figuratively.
The question is what do I mean by literal humour versus figurative
humour? :-)

I also made some typo fixes.

The formatting refers to making the line endings (length) more
consistent with previous settings. In particular in .vimrc:

        autocmd! Filetype markdown setlocal textwidth=80 formatoptions=jt spell spelllang=en_gb ai

though of course one could change the language if they're so inclined.

Incidentally, if it helps, to add spellchecking to git commits as well
as the 51 column for summary and 72 for other lines recommendation, one
can do e.g.:

    hi! gitcommitoverflow ctermbg=black ctermfg=grey
    autocmd! Filetype gitcommit setlocal textwidth=72 formatoptions=jt spell spelllang=en_gb noai nosmartindent

though one could change the colouring scheme and language of course and
obviously some lines have to be longer (try 'J' in command mode to merge
lines where necessary as typing with this set up will cause the lines to
end prematurely in these cases). Of course this is just for those who
use vim but it can help with spelling errors (grammar not so much or
actually at all but grammar checkers are not good at all: for instance
often Apple grammar checkers incorrectly suggests 'your' instead of
'you're' both on mobile OSes and macOS).