Make SYNOPSIS consistent with usage strings
The following man pages were updated:

    	chkentry.1
	iocccsize.1
	mkiocccentry.1
	verge.1

The consistency was checked via a script that I wrote that will most
likely be committed in time (with its own man page of course) but which
needs to be worked on a bit more before this can happen.

This commit also fixed an outdated mkiocccentry man page as it referred
still to the jinfochk and jauthchk tools which no longer are part of the
toolkit being replaced with chkentry.

There might have been some other changes as well.