Add run_usage.sh to check SYNOPSIS consistencies
This script is not perfect. It's a hack. It was quickly written. It's
dirty and ugly but it is helpful in finding inconsistencies in the man
pages wrt the tools. In particular it was used with commit
d7f165c187de5e4847f75490e52505f062526e4c.

There's no man page right now because currently it uses the entire name
of the tool for the man page which means that for shell scripts it would
want e.g. run_usage.sh.1 when it should be run_usage.1. This will be
fixed in another commit another day.

Updated Makefile. Passes shellcheck.