Add run_usage.1 + minor improvements to run_usage
The script was used to generate the correct text to put into run_usage.1
and then verify it works okay (along with a visual inspection).

New option -M to specify the man page file. This is not required but
will be needed when the man page filename should not have extensions of
the tool e.g. run_usage.sh man page is run_usage.1 so -M run_usage.1 is
used there.

If a man page is missing still show the SYNOPSIS (as determined) that
can be added once the man page has been created.

I noted some limitations in the BUGS section including some additional
issues I've noticed. Again this is not perfect by any stretch but for
the most part it does work and will be useful whenever a tool usage
changes. A known issue is that if there's more than one usage string the
script will report both to add but if one is correct in the man page it
can miss the others.