Bug fixes and minor improvements to run_usage.sh
Better handling of missing usage in man page. It's still not perfect but
it at least now checks the number of matches in the man page. This can
still not work properly but fixing it would not be worth it when one has
to still edit the man page and verify it renders okay. This change is
worth it though.

If there's no usage string (either because the tool does not have a -h
option or the -h option does not have any line with 'usage:') the script
no longer shows empty (and thus invalid) possible SYNOPSIS to add.

Moved the limitations comments to the man page (some of which were
already there and some which were not).

Script now says it's possible that there's a missing or inconsistent
SYNOPSIS.

Updated the man page.